WIP import csv genoerator classes and attempt to bring in dependencies for generating csvs.
This is going to take a bit more work to figure out the configuraiton.
Until now, I have sort of been straddling a line between this gem being
able to be run independnetly in an irb session, and requiring Rails modules.
Since the new feature requires rendering (and later features require Rails generators),
we really do need Rails modules. I have tried in this WIP to bring in some of what
we need, but have run into a bunch of partially incomplete dependency trees.
Needless to say, this is not the kind of project I want to dive into with a
limited timeframe, so I am pushing it up as it is now. But I think it may be
something to think more about — do we want to assume that Rails modules
are available to us? Or do we want to try to bring in SOME Rails code
and let this run independently of Rails? (I lean towards the former,
at least as a starting point);